Allow the backend to have empty max_period for resource
A new empty choice field is added to backend for max reservable
period for resource. If this field (max_period) is empty in
backend, then user should be able to choose as many available
slots as they can.

Also, move the getSlotCount function to utils so that it can be
tested.

Refs TAM-107