Add whole opening hours reservation
Resources can have different amount of opening hours for
different days. The users should be able to reserve the resource
for the entire opening hours for a day. E.g. If a resource is
available for 6 hours on Monday and 8 hours on Tuesday. Then
when user clicks to reserve on Monday, it should reserve the
resource for 6 hours , and on Tuesday it should reserve for
8 hours.

Refs TAM-130