Remove the refund policy text for now
The refund policy text needs to be hidden for now. We might want it
again in the future. The refund policy texts are replaced with
empty string so they won't render. The empty string seems to be
easier solution to hide them because other changes made the CI
fail due to test coverage dropping.

Refs TAM-149