Reservations: show correct steps depending on pricing info
This should show only 2 steps not only if resource is free or user can bypass
payment, but also if the total price based on selected pricing list
options is zero.

To make this work, it was necessary to hoist props/state and related
methods from the ReservationInformation component to the parent ReservationPage
component.

Refs TTVA-68