Show payment link button in user reservations
If "payment_link" is not null, then show the link in a "Pay" button in
the user reservations list.

The payment link logic is handled by Respa API, so it should not be
shown e.g. for reservations that are confirmed or cancelled.

The payment link should also not be shown when manual confirmation is
required or the user is admin of the resource/unit. This logic might
need to be revisited when we allow normal payments for staff members.

Refs TTVA-81