Set reservation state to waiting_for_payment when approving paid reservations
The payment return url is also included in the PUT request to enable
payment.

In addition, the "Pay" link has been added to the available actions in
the "My Reservations" page for reservations requiring manual
confirmation, and payment options are likewise enabled for paid
resources in the reservation form (currently these fields are hidden for
manual confirmation resources). This is to enable testing of the full
payment cycle.

Refs TTVA-83