Only show Excel/CSV download controls in Manage Reservations page once totalCount is available
As we use totalCount for the page download, this will be zero while page
is loading the reservations. The user might click prematurely on the "Download" button,
resulting in only 20 rows (server default) being returned.

This change hides the download controls if totalCount is zero i.e. no
reservations found (in which case the Download button will just download
an empty file) or the reservations are still being loaded.

Refs TTVA-90