Resources: check canOnlyBeReservedExternally property to determine whether to show login link
We already check whether resource is reservable, however the canOnlyBeReservedExternally is also set if the resource cannot be reserved due to the resource not having a current bookable period.

This commit also fixes an error in the ResourceInfo component where a
looped element does not have the "key" attribute set.

Refs TTVA-123