Reservations: skip redirect to payment URL if invoice requested
Currently if a payment URL is returned with an order we immediately
redirect to that location (i.e. Ceepos payment page). If an invoice is
requested, instead just redirect to Varaamo reservation confirmation
page.

Also removes the "Pay" step in the reservation steps component, as this
step is handled outside of Varaamo.

Refs: TTVA-142