Add user email address form modal
If the user uses an authentication method that doesn't
provide the an email address, such as PIKI or Suomi.fi,
the user will not have an email address set in Respa.

If that's the case, show a modal with a form for setting
the email address to the user's profile in Respa.

Refs STAM-5