Add a cancel button to the email confirmation form
When the user logs in and doesn't have an email address
saved to the user profile in Respa, a pop-up form is shown
so the user can save an email address to their profile.
Add a cancel button to the pop-up. When clicked, log the
user out. Tunnistamo should then redirect the user back
to the login page.

Refs TTVA-215